:truck: Move function into utils
